Separated components to better improvements react-query
