+++ src/components/AddComment/index.ts
@@ -1,0 +1,2 @@
+export { AddComment } from "./AddComment";
+export type { Values } from "./types";
