+++ src/components/AddComment/types.ts
@@ -1,0 +1,4 @@
+export type Values = {
+    comment: string;
+    files: Array<{ file: File }>,
+};
