+++ src/components/LogIn/index.ts
@@ -1,0 +1,1 @@
+export { LogIn } from "./LogIn";
