+++ src/components/ViewCard/ViewCard.tsx
@@ -1,31 +1,30 @@
-import { FC } from "react";
+import get from "lodash/get";
+import size from "lodash/size";
 import styled from "styled-components";
 import { faCalendarDays } from "@fortawesome/free-solid-svg-icons";
 import {
-    H1,
-    H3,
-    P5,
-    Pill,
-    Icon,
-    Stack,
-    Checkbox,
+    Title,
     HorizontalDivider,
-    useDeskproAppTheme
+    useDeskproAppTheme,
 } from "@deskpro/app-sdk";
+import { H1, H3, P5, Pill, Icon, Stack, Checkbox } from "@deskpro/deskpro-ui";
 import { CardType, ChecklistItem, Comment } from "../../services/trello/types";
 import { getDate } from "../../utils/date";
 import { getLabelColor } from "../../utils";
 import {
+    NoFound,
     LinkIcon,
     EmptyInlineBlock,
     TextBlockWithLabel,
 } from "../common";
 import { Members } from "../common/Cards";
 import { Comments } from "./Comments";
+import type { FC } from "react";
 
-type Props = CardType & {
+type Props =  {
+    card: CardType,
     comments?: Comment[],
-    onAddNewCommentPage: (cardId: CardType["id"]) => void,
+    onNavigateToAddNewComment: () => void,
     onChangeChecklistItem: (
         itemId: ChecklistItem["id"],
         state: ChecklistItem["state"],
@@ -37,40 +36,39 @@ const Description = styled(P5)`
 `;
 
 const ViewCard: FC<Props> = ({
-    id,
-    due,
-    name,
-    desc,
-    list,
-    board,
-    labels,
-    members,
+    card,
     comments,
-    checklists,
-    onAddNewCommentPage,
+    onNavigateToAddNewComment,
     onChangeChecklistItem,
 }) => {
     const { theme } = useDeskproAppTheme();
+    const labels = get(card, ["labels"]);
+    const due = get(card, ["due"]);
+    const checklists = get(card, ["checklists"]);
+
+    if (!card) {
+        return (<NoFound/>);
+    }
 
     return (
         <>
-            <H3 style={{ marginBottom: 10 }}>{name}</H3>
+            <Title title={get(card, ["name"], "-")} />
             <TextBlockWithLabel
                 label="Board"
                 text={(
                     <>
-                        <P5 style={{ marginRight: 4 }}>{board.name}</P5>
-                        <LinkIcon size={10} href={board.url}/>
+                        <P5 style={{ marginRight: 4 }}>{get(card, ["board", "name"], "-")}</P5>
+                        <LinkIcon size={10} href={get(card, ["board", "url"], "-")}/>
                     </>
                 )}
             />
             <TextBlockWithLabel
                 label="List"
-                text={list.name}
+                text={get(card, ["list", "name"], "-")}
             />
             <TextBlockWithLabel
                 label="Description"
-                text={<Description>{desc}</Description>}
+                text={<Description>{get(card, ["desc"], "-")}</Description>}
             />
             <TextBlockWithLabel
                 label="Labels"
@@ -104,9 +102,9 @@ const ViewCard: FC<Props> = ({
                 )}
             />
 
-            <Members members={members} />
+            <Members members={get(card, ["members"])} />
 
-            {(Array.isArray(checklists) && checklists.length > 0) && (
+            {(Array.isArray(checklists) && size(checklists)) && (
                 <>
                     <HorizontalDivider style={{ marginBottom: 10 }} />
                     <H1 style={{ marginBottom: 14 }}>Checklist</H1>
@@ -136,7 +134,7 @@ const ViewCard: FC<Props> = ({
 
             <HorizontalDivider style={{ marginBottom: 10 }} />
 
-            <Comments comments={comments} onClickTitleAction={() => onAddNewCommentPage(id)} />
+            <Comments comments={comments} onClickTitleAction={onNavigateToAddNewComment} />
         </>
     );
 };
