+++ src/components/common/InputSearch/index.ts
@@ -1,1 +1,2 @@
 export { InputSearch } from "./InputSearch";
+export type { Props } from "./types";
