+++ src/components/index.ts
@@ -1,0 +1,5 @@
+export { LogIn } from "./LogIn";
+export { LinkCard } from "./LinkCard";
+export { ViewCard } from "./ViewCard";
+export { AddComment } from "./AddComment";
+export { Home } from "./Home";
