+++ src/main.tsx
@@ -5,7 +5,7 @@ import TimeAgo from "javascript-time-ago";
 import en from "javascript-time-ago/locale/en.json";
 import { DeskproAppProvider } from "@deskpro/app-sdk";
 import { StoreProvider } from "./context/StoreProvider";
-import App from "./App";
+import { App } from "./App";
 
 import "iframe-resizer/js/iframeResizer.contentWindow.js";
 import "flatpickr/dist/themes/light.css";
