--- src/pages/AddComment.tsx
@@ -1,117 +1,0 @@
-import { FC } from "react";
-import isEmpty from "lodash/isEmpty";
-import * as yup from "yup";
-import { useFormik } from "formik";
-import { useNavigate, useParams } from "react-router-dom";
-import {
-    Stack,
-    useDeskproElements,
-    useDeskproAppClient,
-} from "@deskpro/app-sdk";
-import { useStore } from "../context/StoreProvider/hooks";
-import { useSetTitle } from "../hooks";
-import {
-    createAttachService,
-    createCardCommentService,
-} from "../services/trello";
-import { Button, Label, TextArea, Attach, Container } from "../components/common";
-
-const validationSchema = yup.object().shape({
-    comment: yup.string(),
-});
-
-const initValues = {
-    comment: "",
-    files: [],
-};
-
-const AddCommentPage: FC = () => {
-    const navigate = useNavigate();
-    const { cardId } = useParams();
-    const [state, dispatch] = useStore();
-    const { client } = useDeskproAppClient();
-    const {
-        handleSubmit,
-        isSubmitting,
-        setFieldValue,
-        getFieldProps,
-    } = useFormik({
-        validationSchema,
-        initialValues: initValues,
-        onSubmit: async (values) => {
-            if (!client || !cardId) {
-                return
-            }
-
-            // eslint-disable-next-line @typescript-eslint/no-explicit-any
-            const promises: Promise<any>[] = [];
-
-            if (!isEmpty(values.comment)) {
-                promises.push(createCardCommentService(client, cardId, values.comment));
-            }
-
-            if (Array.isArray(values.files) && values.files.length) {
-                promises.concat(values.files.map(({ file }) => {
-                    const form = new FormData();
-                    form.append("file", file);
-                    return createAttachService(client, cardId, form);
-                }))
-            }
-
-            await Promise.all(promises)
-                .then(() => navigate(`/view_card/${cardId}`))
-                .catch((error) => {
-                    dispatch({ type: "error", error })
-                });
-        },
-    });
-
-    useSetTitle("Add Comment");
-
-    useDeskproElements(({ clearElements, registerElement }) => {
-        clearElements();
-        registerElement("trelloRefreshButton", { type: "refresh_button" });
-        registerElement("trelloHomeButton", {
-            type: "home_button",
-            payload: { type: "changePage", path: "/home" }
-        });
-    });
-
-    return (
-        <Container>
-            <form onSubmit={handleSubmit}>
-                <Label htmlFor="comment" label="New comment" marginBottom={17}>
-                    <TextArea
-                        minWidth="auto"
-                        placeholder="Enter comment"
-                        {...getFieldProps("comment")}
-                    />
-                </Label>
-
-                <Label label="Attachments">
-                    <Attach
-                        onFiles={(files) => {
-                            setFieldValue("files", files);
-                        }}
-                    />
-                </Label>
-
-                <Stack justify="space-between">
-                    <Button
-                        type="submit"
-                        text="Save"
-                        disabled={isSubmitting}
-                        loading={isSubmitting}
-                    />
-                    <Button
-                        text="Cancel"
-                        intent="tertiary"
-                        onClick={() => navigate(`/view_card/${cardId}`)}
-                    />
-                </Stack>
-            </form>
-        </Container>
-    );
-};
-
-export { AddCommentPage };
