+++ src/pages/AddCommentPage/index.ts
@@ -1,0 +1,1 @@
+export { AddCommentPage } from "./AddCommentPage";
