+++ src/pages/AdminCallbackPage/index.ts
@@ -1,0 +1,1 @@
+export { AdminCallbackPage } from "./AdminCallbackPage";
