+++ src/pages/CreateCardPage/index.ts
@@ -1,0 +1,1 @@
+export { CreateCardPage } from "./CreateCardPage";
