+++ src/pages/EditCardPage/index.ts
@@ -1,0 +1,1 @@
+export { EditCardPage } from "./EditCardPage";
