+++ src/pages/LoadingAppPage/index.ts
@@ -1,0 +1,1 @@
+export { LoadingAppPage } from "./LoadingAppPage";
