+++ src/pages/ViewCardPage/index.ts
@@ -1,0 +1,1 @@
+export { ViewCardPage } from "./ViewCardPage";
