+++ src/pages/index.ts
@@ -1,0 +1,9 @@
+export { AddCommentPage } from "./AddCommentPage";
+export { AdminCallbackPage } from "./AdminCallbackPage";
+export { CreateCardPage } from "./CreateCardPage";
+export { EditCardPage } from "./EditCardPage";
+export { HomePage } from "./HomePage";
+export { LinkCardPage } from "./LinkCardPage";
+export { LoadingAppPage } from "./LoadingAppPage";
+export { LogInPage } from "./LogInPage";
+export { ViewCardPage } from "./ViewCardPage";
