+++ src/types.ts
@@ -1,10 +1,17 @@
 import type { To, ParamKeyValuePair } from "react-router-dom";
-import type { Context, IDeskproClient } from "@deskpro/app-sdk";
+import type {
+    Context,
+    IDeskproClient,
+    DropdownValueType,
+} from "@deskpro/app-sdk";
 import type { CardType } from "./services/trello/types";
 
+/** Common types */
 export type Maybe<T> = T | undefined | null;
 
 export type Dict<T> = Record<string, T>;
+
+export type Option<Value = unknown> = Omit<DropdownValueType<Value>, "subItems">;
 
 /**
  * An ISO-8601 encoded UTC date time string. Example value: `""2019-09-07T15:50:00Z"`.
