+++ src/utils/getOption.ts
@@ -1,0 +1,16 @@
+import type { DropdownValueType } from "@deskpro/app-sdk";
+import type { Option } from "../types";
+
+const getOption = <Value>(
+  value: Value,
+  label: DropdownValueType<Value>["label"],
+  description?: DropdownValueType<Value>["description"],
+): Option<Value> => ({
+  label,
+  value,
+  key: `${value}`,
+  type: "value",
+  ...(description ? { description } : {}),
+});
+
+export { getOption };
