+++ src/utils/index.ts
@@ -5,3 +5,4 @@ export { getFilteredCards } from "./getFilteredCards";
 export { getEntityMetadata } from "./getEntityMetadata";
 export { getUrlOrigin } from "./getUrlOrigin";
 export { isNavigatePayload } from "./isNavigatePayload";
+export { getOption } from "./getOption";
